+++ Source/Gapotchenko.GnuTK/Toolkits/Cygwin/CygwinToolkit.cs
@@ -27,16 +27,27 @@ sealed class CygwinToolkit(CygwinToolkitFamily family, ICygwinSetupInstance setu
 
     public int ExecuteFile(string path, IReadOnlyList<string> arguments, IReadOnlyDictionary<string, string?>? environment, ToolkitExecutionOptions options)
     {
-        return ExecuteCommand("sh \"$0\" \"$@\"", [path, .. arguments], environment, options);
+        return ExecuteCommandCore("sh \"$0\" \"$@\"", [path, .. arguments], environment, options, null);
     }
 
     public int ExecuteCommand(string command, IReadOnlyList<string> arguments, IReadOnlyDictionary<string, string?>? environment, ToolkitExecutionOptions options)
     {
-        string processPath = GetShellPath();
+        // The 'sh' shell of Cygwin is 'bash' in disguise.
+        return ExecuteCommandCore(command, arguments, environment, options, ["-e", "-o", "pipefail"]);
+    }
+
+    int ExecuteCommandCore(
+        string command,
+        IReadOnlyList<string> commandArguments,
+        IReadOnlyDictionary<string, string?>? environment,
+        ToolkitExecutionOptions options,
+        IEnumerable<string>? extraShellArguments)
+    {
+        string shellPath = GetShellPath();
 
         var psi = new ProcessStartInfo
         {
-            FileName = processPath
+            FileName = shellPath
         };
 
         var processEnvironment = psi.Environment;
@@ -52,14 +63,16 @@ sealed class CygwinToolkit(CygwinToolkitFamily family, ICygwinSetupInstance setu
         {
             // POSIX-compliant behavior requires us to add a lookup path for the shell directory.
             // Otherwise, the files in the shell directory cannot be found by the shell.
-            EnvironmentServices.PrependPath(processEnvironment, Path.GetDirectoryName(processPath));
+            EnvironmentServices.PrependPath(processEnvironment, Path.GetDirectoryName(shellPath));
         }
 
-        var processArguments = psi.ArgumentList;
+        var shellArguments = psi.ArgumentList;
         if (posixifyShell && !posixlyCorrect)
-            processArguments.Add("-posix");
-        processArguments.Add("-l");
-        processArguments.Add("-c");
+            shellArguments.Add("-posix");
+        shellArguments.Add("-l");
+        if (extraShellArguments != null)
+            shellArguments.AddRange(extraShellArguments);
+        shellArguments.Add("-c");
 
         var commandBuilder = new StringBuilder();
         if (!posixlyCorrect)
@@ -69,13 +82,13 @@ sealed class CygwinToolkit(CygwinToolkitFamily family, ICygwinSetupInstance setu
             commandBuilder.Append("unset POSIXLY_CORRECT;");
         }
         commandBuilder.Append(command);
-        processArguments.Add(commandBuilder.ToString());
+        shellArguments.Add(commandBuilder.ToString());
 
-        processArguments.Add(Directory.GetCurrentDirectory());
-        if (arguments is [])
-            processArguments.Add(processPath);
+        shellArguments.Add(Directory.GetCurrentDirectory());
+        if (commandArguments is [])
+            shellArguments.Add(shellPath);
         else
-            processArguments.AddRange(arguments);
+            shellArguments.AddRange(commandArguments);
 
         return ToolkitKit.ExecuteProcess(psi);
     }